Added the undefined tab to the glossary prowindow
