+++ VSProjects/Geomapmaker/Geomapmaker.csproj
@@ -220,6 +220,7 @@
     <Compile Include="TestButton.cs" />
     <Compile Include="ViewModels\Export\ExportVM.cs" />
     <Compile Include="ViewModels\GeneratePolygons\GeneratePolygonsViewModel.cs" />
+    <Compile Include="ViewModels\Glossary\UndefinedGlossaryVM.cs" />
     <Compile Include="ViewModels\Glossary\GlossaryVM.cs" />
     <Compile Include="ViewModels\MapUnitPolysCreate\MapUnitPolysCreateVM.cs" />
     <Compile Include="ViewModels\MapUnitPolysEdit\MapUnitPolysEditVM.cs" />
@@ -228,6 +229,9 @@
     </Compile>
     <Compile Include="Views\Glossary\Glossary.xaml.cs">
       <DependentUpon>Glossary.xaml</DependentUpon>
+    </Compile>
+    <Compile Include="Views\Glossary\UndefinedGlossary.xaml.cs">
+      <DependentUpon>UndefinedGlossary.xaml</DependentUpon>
     </Compile>
     <Compile Include="Views\MapUnitPolysCreate\MapUnitPolysCreate.xaml.cs">
       <DependentUpon>MapUnitPolysCreate.xaml</DependentUpon>
@@ -372,6 +376,10 @@
       <Generator>MSBuild:Compile</Generator>
     </Page>
     <Page Include="Views\Glossary\Glossary.xaml">
+      <Generator>MSBuild:Compile</Generator>
+      <SubType>Designer</SubType>
+    </Page>
+    <Page Include="Views\Glossary\UndefinedGlossary.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
