+++ VSProjects/Geomapmaker/Views/Glossary/Glossary.xaml
@@ -4,9 +4,10 @@
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:extensions="clr-namespace:ArcGIS.Desktop.Extensions;assembly=ArcGIS.Desktop.Extensions"
+        xmlns:extensions="clr-namespace:ArcGIS.Desktop.Extensions;assembly=ArcGIS.Desktop.Extensions" 
+        xmlns:tabData="clr-namespace:Geomapmaker.Views.Glossary"
         mc:Ignorable="d"
-        Title="Update Glossary"
+        Title="Glossary"
         Height="400" 
         MinHeight="350" 
         Width="400"
@@ -18,61 +19,29 @@
         <ResourceDictionary>
             <ResourceDictionary.MergedDictionaries>
                 <extensions:DesignOnlyResourceDictionary Source="pack://application:,,,/ArcGIS.Desktop.Framework;component\Themes\Default.xaml"/>
-                <ResourceDictionary Source="/Geomapmaker;component/Views/GlobalResourceDictionary.xaml" />
             </ResourceDictionary.MergedDictionaries>
         </ResourceDictionary>
     </controls:ProWindow.Resources>
 
     <Grid>
         <Grid.RowDefinitions>
-            <RowDefinition Height="5"/>
-            <RowDefinition Height="Auto"/>
-            <RowDefinition Height="Auto"/>
-            <RowDefinition Height="Auto"/>
-            <RowDefinition Height="Auto"/>
-            <RowDefinition Height="Auto"/>
             <RowDefinition Height="*"/>
-            <RowDefinition Height="Auto"/>
-            <RowDefinition Height="Auto"/>
-            <RowDefinition Height="5"/>
         </Grid.RowDefinitions>
         <Grid.ColumnDefinitions>
-            <ColumnDefinition Width="5"/>
             <ColumnDefinition Width="*" />
-            <ColumnDefinition Width="5"/>
         </Grid.ColumnDefinitions>
 
-        <TextBlock Grid.Row="1" Grid.Column="1" Text="{Binding Heading}" FontSize="18" HorizontalAlignment="Center"/>
-
-        <StackPanel Grid.Row="2" Grid.Column="1" >
-            <TextBlock Text="Dataset" />
-            <TextBox Text="{Binding Dataset}" IsReadOnly="True" IsEnabled="False" />
-        </StackPanel>
-
-        <StackPanel Grid.Row="3" Grid.Column="1" >
-            <TextBlock Text="Field" />
-            <TextBox Text="{Binding Field}" IsReadOnly="True" IsEnabled="False" />
-        </StackPanel>
-
-        <StackPanel Grid.Row="4" Grid.Column="1" >
-            <TextBlock Text="Term" />
-            <TextBox Text="{Binding Term}" IsReadOnly="True" IsEnabled="False" />
-        </StackPanel>
-
-        <TextBlock Grid.Row="5" Grid.Column="1" Text="Definition" />
-        <TextBox Grid.Row="6" Grid.Column="1" Text="{Binding Definition, UpdateSourceTrigger=PropertyChanged}" MinLines="3" />
-
-        <StackPanel Grid.Row="7" Grid.Column="1" >
-            <TextBlock Text="DefinitionSourceID" />
-            <TextBox Text="{Binding DefinitionSourceID}" IsReadOnly="True" IsEnabled="False" />
-        </StackPanel>
-
-        <StackPanel Grid.Row="8" Grid.Column="1" Orientation="Horizontal" HorizontalAlignment="Center" >
-            <Button Padding="5" Command="{Binding CommandCancel}" Style="{DynamicResource Esri_Button}" >Cancel</Button>
-            <Button Padding="5" Command="{Binding CommandSkip}" Style="{DynamicResource Esri_Button}" >Skip</Button>
-            <Button Padding="5" Command="{Binding CommandSave}" Style="{DynamicResource Esri_Button}" >Save</Button>
-        </StackPanel>
-
+        <TabControl Grid.Row="0" Grid.Column="0">
+            <TabItem Header="Undefined" >
+                <tabData:UndefinedGlossary />
+            </TabItem>
+            <!--<TabItem Header="Edit">
+                <tabData:EditDataSource />
+            </TabItem>
+            <TabItem Header="Delete">
+                <tabData:DeleteDataSource />
+            </TabItem>-->
+        </TabControl>
     </Grid>
 
 </controls:ProWindow>
