+++ VSProjects/Geomapmaker/Views/Glossary/UndefinedGlossary.xaml.cs
@@ -1,0 +1,15 @@
+﻿using System.Windows.Controls;
+
+namespace Geomapmaker.Views.Glossary
+{
+    /// <summary>
+    /// Interaction logic for Glossary.xaml
+    /// </summary>
+    public partial class UndefinedGlossary : UserControl
+    {
+        public UndefinedGlossary()
+        {
+            InitializeComponent();
+        }
+    }
+}
